Updated setting quantity in cart

+++ online-store/src/main/java/com/skwarek/onlineStore/data/model/order/Item.java
@@ -2,8 +2,6 @@ package com.skwarek.onlineStore.data.model.order;
 
 import com.skwarek.onlineStore.data.entity.product.Product;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 import java.math.BigDecimal;
 
 /**
@@ -13,9 +11,7 @@ public class Item {
 
     private Product product;
 
-//    @Min(value = 0, message = "{Item.quantity.validation.min}")
-//    @NotNull(message = "{Item.quantity.validation.notNull}")
-    private Integer quantity;
+    private Long quantity;
 
     private BigDecimal itemTotalPrice;
 
@@ -31,11 +27,11 @@ public class Item {
         this.product = product;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(long quantity) {
         this.quantity = quantity;
         this.updateItemTotalPrice();
     }
